Agregar coleccion postman

+++ src/main/java/com/demo/clienteapireactive/repository/ClienteRepository.java
@@ -5,7 +5,7 @@ import com.demo.clienteapireactive.model.Cliente;
 
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
-public interface ClienteRepository extends ReactiveMongoRepository<Cliente,Long> {
+public interface ClienteRepository extends ReactiveMongoRepository<Cliente,Integer> {
